Implemented a JSON parser.
This commit implements an LL(1) lexer and the parser for the JSON specification.
Both the lexer and the parser are integrated in the same structure for the
sake of simplicity.

There is no buffering of tokens. Any token produced by the lexer is immediately
consumed by the parser. In fact, the parser triggers the lexer to create a token
only when it requires one.