Implemented a destructor for JSON values.
The destructor recursively traverses the object tree releasing memory accordingly.
Since, key-value pairs are stored in a hash map, the key needs to persist longer
than the lifetime of a token, which is where we derive the key from. As of now,
tokens are not freed. In the future, we need to implement a recyclable token
that is destroyed along with the parser.

Removed the `TokenChannel` enumeration because the tokens are always generated on
the default channel, given we don't generate tokens for whitespace.